+++ packages/button/button.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script>
-import Icon from '@self-ui/icon';
+import Icon from '@selfui/icon';
 
 export default {
   name: 'SelfButton',