+++ src/components/font-table/index.vue
@@ -8,8 +8,8 @@
 </template>
 
 <script>
-import SelfTable from '@self-ui/table';
-import SelfTableColumn from '@self-ui/table-column';
+import SelfTable from '@selfui/table';
+import SelfTableColumn from '@selfui/table-column';
 
 export default {
   name: 'FontTable',