+++ src/components/icon-set/index.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script>
-import Icon from '@self-ui/icon';
+import Icon from '@selfui/icon';
 
 export default {
   name: 'IconSet',