+++ src/views/table/index.vue
@@ -15,8 +15,8 @@
 </template>
 
 <script>
-import SelfTable from '@self-ui/table';
-import SelfTableColumn from '@self-ui/table-column';
+import SelfTable from '@selfui/table';
+import SelfTableColumn from '@selfui/table-column';
 import PropTable from '@/components/prop-table';
 
 export default {