+++ packages/button/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @self-ui/button
+# @selfui/button
 
 ## 1.0.1
 
@@ -6,4 +6,4 @@
 
 - 730b0a8: Initial release
 - Updated dependencies [730b0a8]
-  - @self-ui/icon@1.0.1
+  - @selfui/icon@1.0.1