+++ packages/dropdown/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @self-ui/dropdown
+# @selfui/dropdown
 
 ## 1.0.1
 
@@ -6,5 +6,5 @@
 
 - 730b0a8: Initial release
 - Updated dependencies [730b0a8]
-  - @self-ui/directives@1.0.1
-  - @self-ui/utils@1.0.1
+  - @selfui/directives@1.0.1
+  - @selfui/utils@1.0.1