+++ packages/mixins/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @self-ui/mixins
+# @selfui/mixins
 
 ## 1.0.1